Generated by Spark: Instead on having a sign up button I was thinking everything could be detection base so for example if user already exist then it will take them to where they enter their password. If user is new the email section will be filled out on the users pervious input and they'll have to do the sign up process like normal
but since we are only doing html and css and this is for only testing and see how the whole preview will look like we need to make some filler 2 filler creditnals.

account@example.com
new@example.com
discord@example.com
account@gmail.com

and also I was thinking let's make it so that if a user and existing discord auth within the email input
that it will redirect them to the authizoration button make the rest of the logic apply

and anything that i'm missing that would make logically sense during this whole flow implement it. so it can be a full functional process.